✨typescript in server
