--- server/index.js
@@ -1,3 +1,0 @@
-// eslint-disable-next-line no-global-assign
-require = require('esm')(module)
-module.exports = require('./server.js')
